azure/az-agent/requirements.txt
python-dotenv 
azure-identity
